chore(settings): delete legacy AppSettingsMenu and relocate source sections
