+++ src/components/SettingsV2/sections/SourcesSection.tsx
@@ -1,19 +1,18 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import { MusicSourcesSection, NativeQueueSyncSection } from '@/components/AppSettingsMenu/SourcesSections';
+import { MusicSourcesSection, NativeQueueSyncSection } from './MusicSourcesSection';
 
 /**
- * SettingsV2 Sources section — phase 2 port (#1450).
+ * SettingsV2 Sources section.
  *
- * Reuses the legacy `MusicSourcesSection` + `NativeQueueSyncSection` directly
- * so v2 reads the same hooks (`useProviderContext`, `useTrackListContext`) and
- * writes the same `STORAGE_KEYS.SPOTIFY_QUEUE_SYNC` /
- * `STORAGE_KEYS.SPOTIFY_QUEUE_CROSS_PROVIDER` keys. Both legacy components
- * self-gate (Music Sources requires ≥2 providers; Native Queue Sync requires
- * `capabilities.hasNativeQueueSync` on a connected provider) and return
- * `null` when their conditions aren't met — the v2 section then renders
- * empty, matching v1 behaviour.
+ * Composes `MusicSourcesSection` + `NativeQueueSyncSection`, which read from
+ * `useProviderContext` / `useTrackListContext` and write the same
+ * `STORAGE_KEYS.SPOTIFY_QUEUE_SYNC` / `STORAGE_KEYS.SPOTIFY_QUEUE_CROSS_PROVIDER`
+ * keys. Both components self-gate (Music Sources requires ≥2 providers; Native
+ * Queue Sync requires `capabilities.hasNativeQueueSync` on a connected
+ * provider) and return `null` when their conditions aren't met — the section
+ * then renders empty.
  */
 
 const SourcesContainer = styled.div`
